requirements.txt
future>=0.16.0
Django==2.0.4
Celery==4.1.0
requests==2.18.4
redis==2.10.6
geopy==1.13.0
shapely==1.6.4
psycopg2-binary==2.7.4
django-allauth==0.35.0
django-cors-headers==2.2.0
django-filter==1.1.0
django-fsm==2.6.0
django-health-check==3.5.1
django-rest-swagger==2.1.2
django-sendgrid-v5==0.6.87
django-storages==1.6.6
django-widget-tweaks==1.4.1
djangorestframework==3.8.2
djangorestframework-jwt==1.11.0
coreapi==2.3.3
gitpython>=2.1.9
google-cloud==0.32.0
pyyaml~=3.12
pytz==2018.4
pymavlink==2.2.8
pyulog==0.6.0
cuid.py==0.2
sendgrid==5.3.0
lxml~=4.2.1
